+++ backend/src/main/java/com/example/shoppingMall/Orders/model/OrdersProducts.java
@@ -5,6 +5,6 @@ import lombok.Data;
 public class OrdersProducts {
     private int id; // 주문 상품 정보
     private String order_id; // 주문 정보
-    private double price; // 상품 금액
+    private int price; // 상품 금액
     private int quantity; // 상품 수량
 }
